+++ docs/README.md
@@ -60,19 +60,10 @@
 - [x] 리팩토링 하고 당첨번호, 보너스 번호에도 소수점 처리하자
 - [ ] 중복된 메소드가 있는지 확인
 
-- [ ] MVC 도입하자. 클래스가 많으니 빠르게 찾기 힘들다.
-    - [x] 일단 집어넣기
-    - [x] view 수정
-    - [ ] controller 수정
-    - [ ] model 수정
-    - [ ] Model은 Controller와 View에 의존하지 않아야 한다.
-    - [ ] View에는 Model에만 의존해야 하고 Controller에는 의존하면 안 된다
-    - [ ] View가 Model로 부터 데이터를 받을 때는, 사용자마다 다르게 보여주어야 하는 데이터에 대해서만 받아야 한다.
-    - [ ] Controller는 Model과 View에 의존해도 된다
-    - [ ] View 가 Model로부터 데이터를 받을 때, 반드시 Controller에서 받아야 한다.
-
 
 - [ ]  이름을 통해 의도를 드러냈는가
+- [ ]  리턴 값, 함수 이름, 파라미터 이름에 의도가 드러났는가(무엇을 하는지 파악할 수 있는가)
+- [ ]  변수 이름에 자료형을 사용하지 않았는가
 - [ ]  기능 요구 사항, 프로그래밍 요구 사항, 과제 진행 요구 사항 항목 잘 지켰는가
 - [ ]  [README.md](http://README.md)
    - [ ]  해당 프로젝트가 어떤 프로젝트인지, 어떤 기능을 담고 있는지 나타냈는가
@@ -80,9 +71,7 @@
    - [ ]  예외적인 상황도 기능 목록에 정리했는가
 - [ ]  값을 하드 코딩 하지 않았는가 - 상수를 만들고 이름을 부여해 변수의 역할이 무엇인지 의도를 드러내자
 - [ ]  구현 순서를 잘 지켰는가(상수 - 인스턴스 변수 - 생성자 - 메서드)
-- [ ]  변수 이름에 자료형을 사용하지 않았는가
 - [ ]  함수가 한 가지 기능만 담당하게 했는가
 - [ ]  한 함수의 길이가 15라인을 넘어가지 않도록 구현했는가
-- [ ]  리턴 값, 함수 이름, 파라미터 이름에 의도가 드러났는가(무엇을 하는지 파악할 수 있는가)
 - [ ]  **의미 없는 주석을 달지 않았는가(테스트 코드도 꼼꼼히 확인)**
